Support SDK install when TMP_DIR is a on a different filesystem
fixes #746
