+++ daml-assistant/daml-project-config/DAML/Project/Util.hs
@@ -3,6 +3,10 @@
 
 module DAML.Project.Util where
 
+import Control.Monad
+import System.Directory.Extra
+import System.FilePath
+
 -- | Same as 'fromRight' but monadic in the applied function.
 fromRightM :: Applicative m => (a -> m b) -> Either a b -> m b
 fromRightM f = either f pure
@@ -10,3 +14,14 @@ fromRightM f = either f pure
 -- | Same as 'fromMaybe' but monadic in the default.
 fromMaybeM :: Applicative m => m a -> Maybe a -> m a
 fromMaybeM d = maybe d pure
+
+copyDirectory :: FilePath -> FilePath -> IO ()
+copyDirectory src target = do
+    files <- listFilesRecursive src
+    forM_ files $ \file -> do
+        let baseName = makeRelative src file
+        let targetFile = target </> baseName
+        createDirectoryIfMissing True (takeDirectory targetFile)
+        copyFile file targetFile
+        p <- getPermissions targetFile
+        setPermissions targetFile p { writable = True }
